add strider
