+++ crabml-core/src/backends/wgpu/mod.rs
@@ -1,2 +1,1 @@
-mod arithmetic;
 mod wgpu_tensor;
